Delete moneyHolindg property
Deleted moneyHolding property in MoneyProcessingUnit
MoneyHolding was representing machine's holding money.
The previous structure assumed that money a customer inserts does not directly go into the money storage but holing the amount of money beforehand(money on transaction) so that machine can return changes from money storage.
The reason behind this structrue is differences between returning changes and inserted money.
For example, when a customer insert 10_000 bill and request 8_000 of change, if machine do not have corresponding change(in Korea, it would be a 5_000 bill and 3 1_000bill), machine will face error.
But by assuming the machine has infinite money(in every form of bills as well), and trying sticking with abstraction paradigm, moneyHolding method looked nothing but needless property.